refactoring on create function
